+++ ShaderLab/Parser/SLParser.h
@@ -3,6 +3,7 @@
 #include <stdio.h>
 #include <stack>
 
+#include "Common/Common.h"
 #include "SLParserData.h"
 
 using namespace shaderlab;
@@ -12,10 +13,10 @@ extern std::stack<SLShaderState*>	g_ShaderStateStack;
 extern const char*                  g_CurrentProgramCode;
 extern ProgramType					g_CurrentProgramType;
 
-void InitLex(const char *c, size_t length);
+void InitLex(const char *c, int32 length);
 
 void FreeLex();
 
 int shaderlabparse();
 
-SLShader* ParseShaderLab(const char* data, size_t length);
+SLShader* ParseShaderLab(const char* data, int32 length);
